+++ Il2CppInterop.Runtime/Runtime/VersionSpecific/Class/Class_29_0.cs
@@ -49,7 +49,7 @@ namespace Il2CppInterop.Runtime.Runtime.VersionSpecific.Class
             public uint initializationExceptionGCHandle;
             public uint cctor_started;
             public uint cctor_finished_or_no_cctor;
-            public ulong cctor_thread;
+            public IntPtr cctor_thread;
             public Il2CppMetadataGenericContainerHandle genericContainerHandle;
             public uint instance_size;
             public uint actualSize;
